chore: Use stable channel
Full relro is the default behaviour on Linux and x64, so the
experimental build flag is not required.

Work around lack of `Vec::into_raw_parts`.

Use a standard `Vec` for node paths. It should provide the same
performance or better than using a `LinkedList`, and it has `retain`
defined which facilitates the `remove` functionality.

Signed-off-by: Graham Brereton <graham.brereton@form3.tech>